+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Fixes
 
+* Fix the problem that the SideMenu could not be opened a second time on IE and Edge.
+
 ## 7.6.2 (2019-07-08)
 
 ### Fixes